[examples] Separate 'image' examples
